pyproject.toml
[tool.poetry]
name = "paperrec"
version = "0.1.0"
description = ""
authors = ["timechess <tyc221@ruc.edu.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
arxiv = "^2.1.3"
prisma = "^0.15.0"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
email-validator = "^2.1.1"
yagmail = "^0.15.293"

[tool.poetry.scripts]
paperrec = "paperrec.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
